# output/final_deliverables/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bluestock Nifty 100 Financial Intelligence API",
    "description": "Institutional-grade REST API providing real-time financial ratios, DuPont drivers, Screener queries, Sector benchmarks, and ReportLab PDF tearsheets for the Nifty 100 universe.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/market-cap/{ticker}": {
      "get": {
        "tags": [
          "Valuation & Market Cap"
        ],
        "summary": "Historical Valuation Multiples & Market Cap (2019\u20132024)",
        "description": "Returns annual valuation multiples (P/E, P/B, EV/EBITDA, Dividend Yield) from 2019 to 2024.\nReturns HTTP 404 if ticker is not found.",
        "operationId": "get_company_valuation_api_v1_market_cap__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Company Valuation Api V1 Market Cap  Ticker  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}